Delete old text_box, replace with alt_text_box
